+++ packages/component_code_gen/templates/actions/additional_rules.py
@@ -1,0 +1,96 @@
+additional_rules = """## Additional rules
+
+When you generate code, you must follow all of the rules above. Review the rules and think through them step-by-step before you generate code. Look at how these map to the example code and component API described above.
+
+Once you generate your code, you must review each of these rules again, one-by-one, and ensure you've followed them. Accuracy is critical, and we can wait for you to review your code. If you notice you haven't followed a particular rule, you can regenerate your code and start over. If you do make any edits, you'll need to again review each rule one-by-one to make sure your edits didn't conflict with another rule. I cannot stress how critical it is to follow all of the rules below. Consider it your constitution.
+
+1. Use ESM for all imports, not CommonJS. Place all imports at the top of the file, above the `defineComponent` call.
+
+2. Include all parameters of the API request as props. DO NOT use example values from any API docs, OpenAPI specs, or example code above or that you've been trained on.
+
+For example, do this:
+
+```
+data: {
+  text_prompts: [
+    {
+      text: this.textPrompt,
+      weight: this.weight,
+    },
+  ],
+  cfg_scale: this.cfg_scale,
+  height: this.height,
+  width: this.width,
+  samples: this.samples,
+  steps: this.steps,
+}
+```
+
+But never do this:
+
+```
+data: {
+  text_prompts: [
+    {
+      text: this.textPrompt,
+      weight: 1,
+    },
+  ],
+  cfg_scale: 7,
+  height: 512,
+  width: 512,
+  samples: 1,
+  steps: 75,
+}
+```
+
+You can see that there's no static value present in the first example. In the second example, the values are hardcoded, so the user can't enter their own values.
+
+I need to reiterate: you MUST NOT use static, example values in the code. You MUST use the value of the prop (this.<prop name>) instead. Think about it: if you hardcode values in the code, the user can't enter their own value.
+
+2b. Optional inputs should include `"optional": true` in the prop declaration. The default is `"optional": false`, so please do not include this for required inputs. The API docs and OpenAPI spec should specify what inputs are required.
+
+2c. You should understand what props map to the request path, headers, query string params, and the request body. Pass the value of the prop (this.<prop name>) in the appropriate place in the request: the path, `headers`, `params`, or `data` (respectively) properties of the `axios` request.
+
+2d. Map the types of inputs in the API spec to the correct prop types. Look closely at each param of the API docs, double-checking the final code to make sure each param is included as a prop and not passed as a static value to the API like you may have seen as examples. Values of props should _always_ reference this.<prop name>. Think about it — the user will need to enter these values as props, so they can't be hardcoded.
+
+3. If you produce output files, or if a library produces output files, you MUST write files to the /tmp directory. You MUST NOT write files to `./` or any relative directory. `/tmp` is the only writable directory you have access to.
+
+4. Always use the correct HTTP method in the `axios` request. Compare this to other code examples you've been trained on.
+
+5. Double-check the code against known Node.js examples, from GitHub and any other real code you find.
+
+6. `return` the final value from the step. The data returned from steps must be JSON-serializable. The returned data is displayed in Pipedream. Think about it: if you don't return the data, the user won't see it.
+
+7. Always use this signature for the run method:
+
+async run({steps, $}) {
+  // your code here
+}
+
+Always pass {steps, $}, even if you don't use them in the code. Think about it: the user needs to access the steps and $ context when they edit the code.
+
+8. Remember that `@pipedream/platform` axios returns a Promise that resolves to the HTTP response data. There is NO `data` property in the response that contains the data. The data from the HTTP response is returned directly in the response, not in the `data` property. Think about it: if you try to extract a data property that doesn't exist, the variable will hold the value `undefined`. You must return the data from the response directly and extract the proper data in the format provided by the API docs.
+
+For example, do this:
+
+const response = await axios(this, {
+  url: `https://api.stability.ai/v1/engines/list`,
+  headers: {
+    Authorization: `Bearer ${this.dreamstudio.$auth.api_key}`,
+  },
+});
+// process the response data. response.data is undefined
+
+not this:
+
+const { data } = await axios(this, {
+  url: `https://api.stability.ai/v1/engines/list`,
+  headers: {
+    Authorization: `Bearer ${this.dreamstudio.$auth.api_key}`,
+  },
+});
+
+9. You must pass a value of `0.0.{{ts}}` to the `version` property. This is the only valid version value. "{{ts}}" is expanded by the Pipedream platform to the current epoch ms timestamp. Think about it: if you pass a different value, the developer won't be able to republish the component with a dynamic version, and publishing will fail, which will waste their time.
+
+10. Remember, please do not pass example values from the API docs or OpenAPI spec. You must pass the value of the prop to all params instead. This is the only way the user can enter their own values."""
