+++ packages/component_code_gen/templates/actions/export_summary.py
@@ -1,0 +1,1 @@
+export_summary = """A short summary should be exported before the end so that the user can quickly read what has happened. This is done by calling `$.export("$summary", "Your summary here")`. This is not optional."""
