+++ packages/component_code_gen/templates/common/app_prop.py
@@ -1,0 +1,15 @@
+app_prop = """This object contains a `props` property, which defines a single prop of type "app":
+
+```javascript
+export default {
+  props: {
+    the_app_name: {
+      type: "app",
+      app: "the_app_name",
+    },
+  }
+  // the rest of the component ...
+})
+```
+
+This lets the user connect their app account to the step, authorizing requests to the app API."""
