+++ packages/component_code_gen/templates/common/auth_example.py
@@ -1,0 +1,3 @@
+auth_example = """Here's an example on how the app authenticates with the API:
+
+{auth_example}"""
