+++ packages/component_code_gen/templates/common/end.py
@@ -1,0 +1,6 @@
+end = """## Remember, return ONLY code
+
+Only return Node.js code. You produce Pipedream component code and ONLY Pipedream component code. DO NOT include any English text before or after the Node.js code. DO NOT say something like "Here's an example..." to preface the code. DO NOT include the code in Markdown code blocks, or format it in any fancy way. Just show me the code.
+
+Consider all the instructions and rules above.
+---"""
