+++ packages/component_code_gen/templates/common/format_instructions.py
@@ -1,0 +1,10 @@
+format_instructions = """Use the following format:
+
+Question: the input question you must answer
+Thought: you should always think about what to do. always escape curly brackets
+Action: the action to take, should be one of [{tool_names}]
+Action Input: the input to the action
+Observation: the result of the action
+... (this Thought/Action/Action Input/Observation can repeat N times)
+Thought: I now know the final answer
+Final Answer: the final answer to the original input question. do not include any other text than the code itself"""
