+++ packages/component_code_gen/templates/common/suffix.py
@@ -1,0 +1,3 @@
+suffix = """Begin!
+Question: {input}
+{agent_scratchpad}"""
