+++ packages/component_code_gen/templates/sources/webhooks/async_run.py
@@ -1,0 +1,1 @@
+async_run = """The run method is called when the component receives an event. The event is passed as the first and only argument to the run method. The event is a JSON object that contains the data from the webhook. The event is emitted by calling `this.$emit`. The first argument to `$emit` is the data to emit. You should only pass relevant data. For example, usually only the event.body is relevant. Headers and others are used to validate the webhook, but shouldn't be emitted. The second argument is an object that contains three fields: `id`, `summary`, and `ts`. The `id` field is a unique identifier for the event. The `summary` field is a human-readable summary of the event. The `ts` field is a timestamp of the event."""
